Diagonal path

+++ src/main/java/xyz/scarabya/astar/domain/Point.java
@@ -25,7 +25,7 @@ public class Point
     private final int hash;
 
     public Point cameFrom;
-    public double fScore;
+    public int fScore;
     public int gScore = 2147483647;
     
     public boolean toInsert = true;
